hoist tag resolution to core
