+++ packages/core/src/resolve.ts
@@ -1,6 +1,6 @@
 import { readFileSync } from 'fs';
 import { getChanges, getExtensionName } from './files';
-
+import { parsePlanFile } from './files/plan/parser';
 import { resolveDependencies } from './deps';
 
 /**
@@ -57,4 +57,70 @@ export const resolveWithPlan = (
   }
 
   return sqlfile.join('\n');
+};
+
+/**
+ * Resolves a tag reference to its corresponding change name.
+ * Tags provide a way to reference specific points in a project's deployment history.
+ * 
+ * @param planPath - Path to the plan file containing tag definitions
+ * @param tagReference - The tag reference to resolve (e.g., "project:@tagName" or "@tagName")
+ * @param currentProject - The current project name (used when tag doesn't specify project)
+ * @returns The resolved change name
+ * @throws Error if tag format is invalid or tag is not found
+ * 
+ * @example
+ * // Resolve a tag in the current project
+ * resolveTagToChangeName('/path/to/launchql.plan', '@v1.0.0', 'myproject')
+ * // Returns: 'schema/v1'
+ * 
+ * @example
+ * // Resolve a tag from another project
+ * resolveTagToChangeName('/path/to/launchql.plan', 'auth:@v2.0.0')
+ * // Returns: 'users/table'
+ */
+export const resolveTagToChangeName = (
+  planPath: string, 
+  tagReference: string, 
+  currentProject?: string
+): string => {
+  // If not a tag reference, return as-is
+  if (!tagReference.includes('@')) {
+    return tagReference;
+  }
+  
+  // Handle simple tag format (@tagName) by prepending current project
+  if (tagReference.startsWith('@') && !tagReference.includes(':')) {
+    if (!currentProject) {
+      const plan = parsePlanFile(planPath);
+      if (!plan.data) {
+        throw new Error(`Could not parse plan file: ${planPath}`);
+      }
+      currentProject = plan.data.project;
+    }
+    tagReference = `${currentProject}:${tagReference}`;
+  }
+  
+  // Parse project:@tagName format
+  const match = tagReference.match(/^([^:]+):@(.+)$/);
+  if (!match) {
+    throw new Error(`Invalid tag format: ${tagReference}. Expected format: project:@tagName or @tagName`);
+  }
+  
+  const [, projectName, tagName] = match;
+  
+  // Parse plan file to find tag
+  const planResult = parsePlanFile(planPath);
+  
+  if (!planResult.data) {
+    throw new Error(`Could not parse plan file: ${planPath}`);
+  }
+  
+  // Find the tag in the plan
+  const tag = planResult.data.tags?.find((t: any) => t.name === tagName);
+  if (!tag) {
+    throw new Error(`Tag ${tagName} not found in project ${projectName}`);
+  }
+  
+  return tag.change;
 };
